requirements.txt
SQLAlchemy==1.1.3
argparse==1.2.1
psycopg2==2.6.2
python-wordpress-xmlrpc==2.3
wsgiref==0.1.2
